+++ ignite/cmd/scaffold.go
@@ -1,8 +1,6 @@
 package ignitecmd
 
 import (
-	"path/filepath"
-
 	"github.com/manifoldco/promptui"
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
@@ -10,7 +8,6 @@ import (
 	"github.com/ignite/cli/v29/ignite/pkg/cliui"
 	"github.com/ignite/cli/v29/ignite/pkg/cosmosver"
 	"github.com/ignite/cli/v29/ignite/pkg/errors"
-	"github.com/ignite/cli/v29/ignite/pkg/gomodulepath"
 	"github.com/ignite/cli/v29/ignite/pkg/xgit"
 	"github.com/ignite/cli/v29/ignite/services/scaffolder"
 	"github.com/ignite/cli/v29/ignite/version"
@@ -23,6 +20,7 @@ const (
 	flagNoSimulation = "no-simulation"
 	flagResponse     = "response"
 	flagDescription  = "desc"
+	flagProtoDir     = "proto-dir"
 
 	msgCommitPrefix = "Your saved project changes have not been committed.\nTo enable reverting to your current state, commit your saved changes."
 	msgCommitPrompt = "Do you want to proceed without committing your saved changes"
@@ -142,13 +140,12 @@ func migrationPreRunHandler(cmd *cobra.Command, args []string) error {
 	session := cliui.New()
 	defer session.End()
 
-	path := flagGetPath(cmd)
-	path, err := filepath.Abs(path)
+	cfg, _, err := getChainConfig(cmd)
 	if err != nil {
 		return err
 	}
 
-	_, appPath, err := gomodulepath.Find(path)
+	appPath, err := goModulePath(cmd)
 	if err != nil {
 		return err
 	}
@@ -166,7 +163,7 @@ func migrationPreRunHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	return bufMigrationPreRunHandler(cmd, session, appPath)
+	return bufMigrationPreRunHandler(cmd, session, appPath, cfg.Build.Proto.Path)
 }
 
 func scaffoldType(
@@ -206,7 +203,12 @@ func scaffoldType(
 	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))
 	defer session.End()
 
-	sc, err := scaffolder.New(cmd.Context(), appPath)
+	cfg, _, err := getChainConfig(cmd)
+	if err != nil {
+		return err
+	}
+
+	sc, err := scaffolder.New(cmd.Context(), appPath, cfg.Build.Proto.Path)
 	if err != nil {
 		return err
 	}