+++ ignite/cmd/scaffold_module.go
@@ -121,6 +121,11 @@ func scaffoldModuleHandler(cmd *cobra.Command, args []string) error {
 	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))
 	defer session.End()
 
+	cfg, _, err := getChainConfig(cmd)
+	if err != nil {
+		return err
+	}
+
 	ibcModule, _ := cmd.Flags().GetBool(flagIBC)
 	ibcOrdering, _ := cmd.Flags().GetString(flagIBCOrdering)
 	requireRegistration, _ := cmd.Flags().GetBool(flagRequireRegistration)
@@ -171,7 +176,7 @@ func scaffoldModuleHandler(cmd *cobra.Command, args []string) error {
 	var msg bytes.Buffer
 	fmt.Fprintf(&msg, "\n🎉 Module created %s.\n\n", name)
 
-	sc, err := scaffolder.New(cmd.Context(), appPath)
+	sc, err := scaffolder.New(cmd.Context(), appPath, cfg.Build.Proto.Path)
 	if err != nil {
 		return err
 	}