+++ ignite/templates/module/create/msgserver.go
@@ -40,6 +40,7 @@ func AddMsgServerConventionToLegacyModule(replacer placeholder.Replacer, opts *M
 
 	plushhelpers.ExtendPlushContext(ctx)
 	g.Transformer(xgenny.Transformer(ctx))
+	g.Transformer(genny.Replace("{{protoDir}}", opts.ProtoDir))
 	g.Transformer(genny.Replace("{{appName}}", opts.AppName))
 	g.Transformer(genny.Replace("{{moduleName}}", opts.ModuleName))
 	return g, nil